requirements.txt
fastapi
uvicorn
line-bot-sdk
google.generativeai